VT100 line wrapping and control character support
This change adds support for VT100-style line wrapping: when the cursor
reaches the end of a line, it does not immediately move to the next line.
Instead, it stays at the end of the line until the next printable
character arrives.

With this change, full-screen applications work much better because they
can successfully write to the lower-right corner of the screen.

This change also fixes support for various C0 control characters.

Signed-off-by: John Starks <jostarks@microsoft.com>