winterm: fix GetStdFile() falltrough
Commit 3ee0038cec469c02f7a130c7824d74a442330f35 updated the switch
with the intent to handle either syscall.XX or golang.org/x/sys/win.XX
values.

However, Go "switch" does not have an implicit "fallthrough" behavior,
and either requires the "fallthrough" keyword to be used, or to combine
values in a single "case".

For example, the following:

    for _, v := range []int{1, 2, 3, 4, 5, 6, 7} {
        result := "unknown"
        switch v {
        case 1: // doesn't fall-through, so this is an empty "case"
        case 2:
            result = "found one or two"
        case 3:
            fallthrough // falls through to "case 4"
        case 4:
            result = "found three or four"
        case 5, 6: // matches either 5 or 6
            result = "found five or six"
        default:
            result = "found something else"
        }
        fmt.Printf("%d: %s\n", v, result)
    }

Will print:

    1: unknown
    2: found one or two
    3: found three or four
    4: found three or four
    5: found five or six
    6: found five or six
    7: found something else

This patch combines "equal" values in a single case within the switch to
fix this issue.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>