Fix OSC string terminator parsing
The OSC-string state handler was incorrectly transitioning to the ground
state on any 0x5C byte. 0x5C is the backslash character `\`, which is a
printable character and valid to include inside of an OSC string. Fix
the parser so that no printable characters are interpreted as OSC
terminators.

Fixes #34

Signed-off-by: Cory Snider <csnider@mirantis.com>